Move `ImportStreamDialog`
